moves VariableByBlankNodeSubstitutionException to the 'se.liu.ida.hefquin.base.query' package
