+++ hefquin-base/src/main/java/se/liu/ida/hefquin/base/query/impl/GenericSPARQLGraphPatternImpl1.java
@@ -21,6 +21,7 @@ import se.liu.ida.hefquin.base.data.SolutionMapping;
 import se.liu.ida.hefquin.base.query.SPARQLGraphPattern;
 import se.liu.ida.hefquin.base.query.SPARQLGroupPattern;
 import se.liu.ida.hefquin.base.query.TriplePattern;
+import se.liu.ida.hefquin.base.query.VariableByBlankNodeSubstitutionException;
 
 /**
  * This class is a generic implementation of {@link SPARQLGraphPattern}
