+++ hefquin-base/src/main/java/se/liu/ida/hefquin/base/query/impl/SPARQLGroupPatternImpl.java
@@ -13,6 +13,7 @@ import se.liu.ida.hefquin.base.data.SolutionMapping;
 import se.liu.ida.hefquin.base.query.SPARQLGraphPattern;
 import se.liu.ida.hefquin.base.query.SPARQLGroupPattern;
 import se.liu.ida.hefquin.base.query.TriplePattern;
+import se.liu.ida.hefquin.base.query.VariableByBlankNodeSubstitutionException;
 
 public class SPARQLGroupPatternImpl implements SPARQLGroupPattern
 {
