+++ hefquin-base/src/main/java/se/liu/ida/hefquin/base/query/impl/TriplePatternImpl.java
@@ -11,6 +11,7 @@ import se.liu.ida.hefquin.base.data.SolutionMapping;
 import se.liu.ida.hefquin.base.query.BGP;
 import se.liu.ida.hefquin.base.query.SPARQLGraphPattern;
 import se.liu.ida.hefquin.base.query.TriplePattern;
+import se.liu.ida.hefquin.base.query.VariableByBlankNodeSubstitutionException;
 
 import java.util.Collections;
 import java.util.HashSet;
